migrate newsletter component
