+++ src/components/footer.tsx
@@ -1,6 +1,7 @@
 import React from "react"
 import { Link } from "gatsby"
 import Text from "./text"
+import Newsletter from "./newsletter"
 
 interface FooterProps {
   siteTitle: string
@@ -57,7 +58,7 @@ export default function Footer({ siteTitle }: FooterProps) {
           <div className="grid grid-cols-2 gap-8 xl:col-span-2">
             <div className="md:grid md:grid-cols-2 md:gap-8">
               <div>
-                <h3 className="text-sm font-semibold text-white tracking-wider uppercase">
+                <h3 className="text-sm font-bold text-white tracking-wider uppercase">
                   À propos
                 </h3>
                 <ul role="list" className="mt-4 space-y-4">
@@ -73,8 +74,8 @@ export default function Footer({ siteTitle }: FooterProps) {
                   ))}
                 </ul>
               </div>
-              <div>
-                <h3 className="text-sm font-semibold text-white tracking-wider uppercase">
+              <div className="mt-6 sm:mt-0 ">
+                <h3 className="text-sm font-bold text-white tracking-wider uppercase">
                   Le podcast
                 </h3>
                 <ul role="list" className="mt-4 space-y-4">
@@ -93,35 +94,7 @@ export default function Footer({ siteTitle }: FooterProps) {
             </div>
           </div>
           <div className="mt-8 xl:mt-0">
-            <h3 className="text-sm font-semibold text-white tracking-wider uppercase">
-              Inscrivez-vous à ma newsletter
-            </h3>
-            <p className="mt-4 text-base text-gray-200">
-              Recevez la liste des podcasts et des articles publiés directement
-              par email.
-            </p>
-            <form className="mt-4 sm:flex sm:max-w-md">
-              <label htmlFor="email-address" className="sr-only">
-                Email address
-              </label>
-              <input
-                type="email"
-                name="email-address"
-                id="email-address"
-                autoComplete="email"
-                required
-                className="appearance-none min-w-0 w-full bg-white border border-gray-300 rounded-md shadow-sm py-2 px-4 text-base text-gray-900 placeholder-gray-500 focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 focus:placeholder-gray-400"
-                placeholder="Enter your email"
-              />
-              <div className="mt-3 rounded-md sm:mt-0 sm:ml-3 sm:flex-shrink-0">
-                <button
-                  type="submit"
-                  className="w-full bg-indigo-600 flex items-center justify-center border border-transparent rounded-md py-2 px-4 text-base font-medium text-white hover:bg-indigo-700 focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500"
-                >
-                  Subscribe
-                </button>
-              </div>
-            </form>
+            <Newsletter />
           </div>
         </div>
         <div className="mt-8 pt-8 md:flex md:items-center md:justify-between">
