+++ src/components/header.tsx
@@ -9,7 +9,7 @@ const HeaderLink = ({ to, title }: { to: string; title: string }) => {
       <Link
         to={to}
         activeClassName="text-lime-700 relative"
-        className="py-4 px-8 hover:text-lime-600 font-extrabold text-lg"
+        className="py-4 px-8 hover:text-black font-extrabold text-lg"
       >
         <span className="relative">{title}</span>
       </Link>
@@ -33,7 +33,7 @@ export default function Header(): ReactElement {
 
         <div className="m-auto">
           <Link
-            className="font-bold text-gray-700 hover:text-lime-600 text-2xl sm:text-5xl pt-6 pb-0 sm:pt-12 flex justify-center transition duration-300 ease-in-out"
+            className="font-bold text-gray-700 hover:text-black text-2xl sm:text-5xl pt-6 pb-0 sm:pt-12 flex justify-center transition duration-300 ease-in-out"
             to={`/`}
           >
             ART <span className="italic tracking-tighter ml-2">au féminin</span>
@@ -50,7 +50,7 @@ export default function Header(): ReactElement {
                   title="Envoyer un email"
                   target="_blank"
                   rel="noopener noreferrer"
-                  className="py-4 px-8 hover:text-lime-600 font-extrabold text-lg"
+                  className="py-4 px-8 hover:text-black font-extrabold text-lg"
                 >
                   Me contacter par email
                 </a>
