+++ src/components/input.stories.tsx
@@ -1,24 +1,21 @@
 import React from "react"
 import { ComponentStory, ComponentMeta } from "@storybook/react"
-import Hr from "../components/hr"
+import Input from "./input"
 
 // More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 export default {
-  title: "Example/Hr",
-  component: Hr,
+  title: "Example/Input",
+  component: Input,
   // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
   argTypes: {
     backgroundColor: { control: "color" },
   },
-} as ComponentMeta<typeof Hr>
+} as ComponentMeta<typeof Input>
 
-// More on component templates: https://storybook.js.org/docs/react/writing-stories/introduction#using-args
-const Template: ComponentStory<typeof Hr> = (args) => {
+export const InputStory = () => {
   return (
     <>
-      <Hr />
+      <Input variant="rounded" />
     </>
   )
 }
-
-export const Primary = Template.bind({})
